Add importlib.resources fallback for VERSION
Adding in this try/except block for finding the version fixes the issue
I was seeing with PyOxidizer. PyOxidizer presents a `files` method, but
it doesn't do anything so you end up with an error when you try to
traverse the returned object.

Now if there are any problems getting the version, we fallback to the
original importlib function. This also has the benefit of supporting
Python 3.7 and up, rather than just 3.9 and up.